+++ app/jobs/create_stats_job.rb
@@ -2,7 +2,7 @@ class CreateStatsJob < ApplicationJob
   def perform
     department_ids = Department.all.collect(&:id).push(nil)
     department_ids.each do |department_id|
-      CreateStat.call(department_id)
+      Stats::CreateStat.call(department_id: department_id)
     end
   end
 end
