--- app/views/stats/_old_stats.html.erb
@@ -1,57 +1,0 @@
-<div class="container marianne">
-  <%= render 'header' %>
-  <div class="row d-flex justify-content-center mt-4 flex-wrap">
-    <div class="col-12 col-md-6 px-5 pb-5">
-      <p class="highlight-stat big margin-left"><%= @stats.applicants.to_a.count %></p>
-      <p class="highlight-stat margin-left">bénéficiaires du RSA<br/>gérés dans RDV-Insertion</p>
-      <%= @stats.applicants.group_by_month(:created_at, format: "%m/%Y", current: false).count %>
-      <%= line_chart @stats.applicants.group_by_month(:created_at, format: "%m/%Y", current: false).count, title: "Nouveaux bénéficiaires par mois", colors: ["#083b66"] %>
-    </div>
-    <div class="col-12 col-md-6 px-5 pb-5 background-blue-light">
-      <p class="highlight-stat big margin-left"><%= @stats.rdvs.to_a.count %></p>
-      <p class="highlight-stat margin-left">rendez-vous pris</p>
-      <%= line_chart @stats.rdvs.group_by_month(:created_at, format: "%m/%Y", current: false).count, title: "Nouveaux rdvs par mois", colors: ["#083b66"] %>
-    </div>
-  </div>
-  <% if current_page?(action: 'index') || @display_all_stats %>
-    <div class="row d-flex justify-content-center flex-wrap-reverse">
-      <div class="col-12 col-md-6 px-5 pb-5 background-blue-light">
-        <p class="highlight-stat big margin-left"><%= @stats.sent_invitations.to_a.count %></p>
-        <p class="highlight-stat margin-left">invitations envoyées<br/>par mail, SMS, et courrier</p>
-        <%= line_chart @stats.sent_invitations.group_by_month(:created_at, format: "%m/%Y", current: false).count, title: "Invitations envoyées par mois", colors: ["#083b66"] %>
-      </div>
-      <div class="col-12 col-md-6 px-5 pb-5">
-        <p class="highlight-stat big margin-left"><%= @stats.percentage_of_no_show.round %> %</p>
-        <p class="highlight-stat margin-left">de rendez-vous<br/>non honorés avec RDV-Insertion</p>
-      <%= line_chart @stats.percentage_of_no_show_by_month, title: "Taux de rdvs non honorés par mois", suffix: "%", colors: ["#083b66"] %>
-      </div>
-    </div>
-    <div class="row d-flex justify-content-center flex-wrap">
-      <div class="col-12 col-md-6 px-5 pb-5">
-        <p class="highlight-stat big margin-left"><%= @stats.average_time_between_invitation_and_rdv_in_days.round %> jours</p>
-        <p class="highlight-stat margin-left">délai moyen entre l'invitation<br/>et la prise de rendez-vous</p>
-        <%= line_chart @stats.average_time_between_invitation_and_rdv_in_days_by_month, title: "Délai moyen d'invitation par mois", ytitle: "jours", colors: ["#083b66"] %>
-      </div>
-      <div class="col-12 col-md-6 px-5 pb-5 background-blue-light">
-        <p class="highlight-stat big margin-left"><%= @stats.average_time_between_rdv_creation_and_start_in_days.round %> jours</p>
-        <p class="highlight-stat margin-left">délai moyen entre la prise<br/>d'un rendez-vous et sa tenue</p>
-        <%= line_chart @stats.average_time_between_rdv_creation_and_start_in_days_by_month, title: "Délai moyen de rdv par mois", ytitle: "jours", colors: ["#083b66"] %>
-      </div>
-    </div>
-    <div class="row d-flex justify-content-center flex-wrap-reverse">
-      <div class="col-12 col-md-6 px-5 pb-5 background-blue-light">
-        <p class="highlight-stat big margin-left"><%= @stats.percentage_of_applicants_oriented_in_less_than_30_days.round %> %</p>
-        <p class="highlight-stat margin-left">de bénéficiaires <strong>orientés en - de 30 jours</strong></p>
-        <%= line_chart @stats.percentage_of_applicants_oriented_in_less_than_30_days_by_month, title: "Taux de bRSA orientés - de 30 jours par mois", suffix: "%", colors: ["#083b66"] %>
-      </div>
-      <div class="col-12 col-md-6 px-5 pb-5 d-flex flex-column justify-content-center align-items-center">
-        <p class="highlight-stat big margin-left"><%= @stats.relevant_agents.to_a.count %> agents</p>
-        <% if current_page?(action: 'index') %>
-          <p class="highlight-stat margin-left">travaillant pour <strong><%= @department_count %> départements</strong><br/> utilisent RDV-Insertion</p>
-        <% else %>
-          <p class="highlight-stat margin-left">utilisent RDV-Insertion</p>
-        <% end %>
-      </div>
-    </div>
-  <% end %>
-</div>
