+++ app/views/stats/_stats.html.erb
@@ -29,19 +29,19 @@
       <div class="col-12 col-md-6 px-5 pb-5">
         <p class="highlight-stat big margin-left"><%= @stat.average_time_between_invitation_and_rdv_in_days.round %> jours</p>
         <p class="highlight-stat margin-left">délai moyen entre l'invitation<br/>et la prise de rendez-vous</p>
-        <%= line_chart @stat.average_time_between_invitation_and_rdv_in_days_grouped_by_month, title: "Délai moyen d'invitation par mois", ytitle: "jours", colors: ["#083b66"] %>
+        <%= line_chart @stat.average_time_between_invitation_and_rdv_in_days_by_month, title: "Délai moyen d'invitation par mois", ytitle: "jours", colors: ["#083b66"] %>
       </div>
       <div class="col-12 col-md-6 px-5 pb-5 background-blue-light">
         <p class="highlight-stat big margin-left"><%= @stat.average_time_between_rdv_creation_and_start_in_days.round %> jours</p>
         <p class="highlight-stat margin-left">délai moyen entre la prise<br/>d'un rendez-vous et sa tenue</p>
-        <%= line_chart @stat.average_time_between_rdv_creation_and_start_in_days_grouped_by_month, title: "Délai moyen de rdv par mois", ytitle: "jours", colors: ["#083b66"] %>
+        <%= line_chart @stat.average_time_between_rdv_creation_and_start_in_days_by_month, title: "Délai moyen de rdv par mois", ytitle: "jours", colors: ["#083b66"] %>
       </div>
     </div>
     <div class="row d-flex justify-content-center flex-wrap-reverse">
       <div class="col-12 col-md-6 px-5 pb-5 background-blue-light">
-        <p class="highlight-stat big margin-left"><%= @stats.percentage_of_applicants_with_rdv_seen_in_less_than_30_days.round %> %</p>
+        <p class="highlight-stat big margin-left"><%= @stat.rate_of_applicants_with_rdv_seen_in_less_than_30_days.round %> %</p>
         <p class="highlight-stat margin-left">de bénéficiaires <strong>ayant honoré un rendez-vous en - de 30 jours</strong></p>
-        <%= line_chart @stats.percentage_of_applicants_with_rdv_seen_in_less_than_30_days_grouped_by_month, title: "Taux de bRSA ayant honoré un rendez-vous en - de 30 jours par mois", suffix: "%", colors: ["#083b66"] %>
+        <%= line_chart @stat.rate_of_applicants_with_rdv_seen_in_less_than_30_days_by_month, title: "Taux de bRSA ayant honoré un rendez-vous en - de 30 jours par mois", suffix: "%", colors: ["#083b66"] %>
       </div>
       <div class="col-12 col-md-6 px-5 pb-5 d-flex flex-column justify-content-center align-items-center">
         <p class="highlight-stat big margin-left"><%= @stat.agents_count %> agents</p>
