+++ config/schedule.yml
@@ -1,0 +1,3 @@
+create_stats_job:
+  cron: "0 22 * * 6" # we compute stats once a week, the saturday at 22:00
+  class: "CreateStatsJob"
