+++ db/migrate/20220614084655_create_stats.rb
@@ -1,6 +1,6 @@
 # rubocop:disable Metrics/AbcSize
 
-class CreateStatsTable < ActiveRecord::Migration[7.0]
+class CreateStats < ActiveRecord::Migration[7.0]
   def change
     create_table :stats do |t|
       t.integer :applicants_count
@@ -12,11 +12,11 @@ class CreateStatsTable < ActiveRecord::Migration[7.0]
       t.float :percentage_of_no_show
       t.json :percentage_of_no_show_grouped_by_month
       t.float :average_time_between_invitation_and_rdv_in_days
-      t.json :average_time_between_invitation_and_rdv_in_days_grouped_by_month
+      t.json :average_time_between_invitation_and_rdv_in_days_by_month
       t.float :average_time_between_rdv_creation_and_start_in_days
-      t.json :average_time_between_rdv_creation_and_start_in_days_grouped_by_month
-      t.float :percentage_of_applicants_oriented_in_less_than_30_days
-      t.json :percentage_of_applicants_oriented_in_less_than_30_days_grouped_by_month
+      t.json :average_time_between_rdv_creation_and_start_in_days_by_month
+      t.float :rate_of_applicants_with_rdv_seen_in_less_than_30_days
+      t.json :rate_of_applicants_with_rdv_seen_in_less_than_30_days_by_month
       t.integer :agents_count
       t.integer :department_id
